+++ package.json
@@ -13,25 +13,22 @@
     "@react-navigation/bottom-tabs": "^7.4.7",
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/native-stack": "^7.3.26",
-    "expo": "^54.0.8",
+    "expo": "^54.0.7",
     "expo-camera": "~17.0.8",
     "expo-secure-store": "~15.0.7",
     "expo-splash-screen": "~31.0.10",
     "expo-status-bar": "~3.0.8",
     "formik": "^2.4.6",
     "react": "19.1.0",
-    "react-dom": "19.1.0",
     "react-native": "0.81.4",
     "react-native-maps": "1.20.1",
     "react-native-safe-area-context": "^5.6.1",
     "react-native-screens": "~4.16.0",
-    "react-native-web": "^0.21.0",
     "yup": "^1.7.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~19.1.10",
-    "expo-module-scripts": "^5.0.7",
     "typescript": "~5.9.2"
   },
   "private": true
